+++ src/index.js
@@ -1,16 +1,13 @@
 import { getOptions } from 'loader-utils';
-import validateOptions from 'schema-utils';
 
-import parse from './parse';
+import validateOptions from 'schema-utils';
+import FrontMatter from 'front-matter';
+import Remark from 'remark';
+import Report from 'vfile-reporter';
+
 import schema from './options.json';
 
-/**
- * Primary loader function
- *
- * @param {string} content - Markdown file content
- */
-export default function loader(content) {
-  const callback = this.async();
+export default function loader(markdown) {
   const options = getOptions(this);
 
   validateOptions(schema, options, {
@@ -18,9 +15,31 @@ export default function loader(content) {
     baseDataPath: 'options',
   });
 
-  parse(content, options)
-    // @todo we should probably just intercept images in the tree
-    // or recommend that the `html-loader` be chained
-    .then((resolved) => callback(null, resolved.content))
-    .catch(callback);
+  const callback = this.async();
+
+  const { plugins = [] } = options;
+  const parsed = FrontMatter(markdown);
+
+  plugins
+    .reduce((remark, item) => {
+      if (Array.isArray(item)) {
+        return remark.use.apply(null, item);
+      }
+
+      return remark.use(item);
+    }, Remark())
+    .process(parsed.body, (err, file) => {
+      const result = {
+        content: file.contents,
+        attributes: parsed.attributes,
+      };
+
+      if (err) {
+        callback(Report(err || file));
+
+        return;
+      }
+
+      callback(null, result.content);
+    });
 }
